requirements.txt
Flask
Flask-SQLAlchemy
Pillow
rarfile